refactor: extract pipVariant to its own module so the unit test can load it
The vitest run failed at the import boundary: pipVariant.test.ts (then
RowPips.test.ts) imports from RowPips.tsx, which transitively pulls
useTerminalStore → wire.ts → solid-sonner — and the test environment
can't resolve 'solid-js/web's 'use' export in that chain. Hickey #4
originally noted the rename was 'the minimum fix' and the extraction was
'also fine, but heavier' — heavier is now forced by the test runner.
Putting the pure dispatch in its own file kills the import chain at the
boundary (pipVariant.ts has zero Solid deps) and the filename matches
the test once more.
