+++ packages/client/src/canvas/dock/RowPips.tsx
@@ -52,6 +52,7 @@ import type { TerminalDisplayInfo } from "../../terminal/terminalDisplay";
 import { useTerminalStore } from "../../terminal/useTerminalStore";
 import { PrStateIcon } from "../../ui/Icons";
 import type { DockRowBucket } from "./dockRowRanking";
+import { type PipVariant, pipVariant } from "./pipVariant";
 import { SubCountChip } from "./SubCountChip";
 
 /** Per-row combined reactive data — `info` + `meta` in a single memo.
@@ -69,37 +70,6 @@ export function createDockRowData(
     if (!info || !meta) return null;
     return { info, meta };
   });
-}
-
-/** State-pip render variant — the dispatch the JSX renders. Pure
- *  function of `(bucket, unread)` so the rule is independently
- *  testable without spinning up a Solid render harness. Order
- *  matters: `unread` dominates the bucket — a row that just fired
- *  an alert in the background still reads as "attention" even if
- *  the agent has since transitioned to working, because the unread
- *  obligation outlives the bucket transition until the user
- *  activates the row. */
-export type PipVariant =
-  | "attention" // unread: loud filled disk + halo + pulse
-  | "awaiting" // bucket awaiting, !unread: quiet dim dot (lingering)
-  | "working" // hollow spinning ring
-  | "idle" // muted small dot
-  | "empty"; // parked / none — render nothing
-
-/** Pure A→B bucket-to-pip mapping. TypeScript's required-property check
- *  on `Record<DockRowBucket, …>` enforces exhaustiveness — adding a new
- *  bucket variant to `DockRowBucket` causes a compile error here. */
-const BUCKET_TO_PIP: Record<DockRowBucket, PipVariant> = {
-  awaiting: "awaiting",
-  working: "working",
-  idle: "idle",
-  parked: "empty",
-  none: "empty",
-};
-
-export function pipVariant(bucket: DockRowBucket, unread: boolean): PipVariant {
-  if (unread) return "attention";
-  return BUCKET_TO_PIP[bucket];
 }
 
 /** Inline PR pip — leading glyph on row line 2. Caller controls
